VGrid deletion:
The vgrid will be deleted when allowing the last owner to leave.
  This is allowed when 
   - it has got no sub-vgrids
   - no local resources participations or local members exist 
       ("local" meaning not inherited, but only for the particular vgrid).  
When deleting, the files of a sub-vgrid will be kept (in directories of the 
parent vgrid), files of top-level VGrids will be simply removed.
As mentioned, sub-vgrids, resources and members have to be removed before deleting.

On the way, refactored/simplified the code to remove vgrid owners.