Changed the VO entity i order to be in line with the sgas definition.
As the VO entrance uses the vo: namespace, i have added this to the XML documents header. 

I have changeed the set_element() function to take namepace and prefix as parameters, i order to support the building of the VO entity. These parameters defaults to namespace_ogf and namespace_prefix, so no changes is needed to the rest of the code.